refactor(workersvc): move claim assembly to claim_assembly.go
Pure same-package code motion (PRD #921 M2, seam 5b). Moves the claim-assembly
cluster (assembleClaim, inflightTargets, formatInflightLine, selfImproveOpenMRs,
firstNonEmptyLine, resolveTooling, loadToolRules, decodePackageList) out of
service.go into the new claim_assembly.go, byte-identically. recoverClaimAssembly
and errToolPackagesRejected stay in service.go (shared). The now-unused forge,
privcheck, toolprofile and toolseed imports are dropped from service.go (pure
deletions). No exported-API change, no behavior change; go doc output unchanged.
service.go is now 3681 lines (was 6508).
